Add policy methods to -becomeFirstResponder & -contentOffset adjustment.
By default, TITokenField call -[self becomeFirstResponder] when a token is added.
This will cause the keyboard to be displayed, which in some cases, may not be
desirable.  The 'becomesFirstResponderWhenAddingToken' flag allows this behavior
to be turned off.

If TITokenField is the child of a UIScrollView, it will adjust the contentOffset
of its parent when -setResultsModeEnabled: is called. This setting can now be
configured by settings 'adjustsScrollView', for cases where clients need to maintain
absolute control of the scroll offset.