Display placeholder text when tokens & text are empty.
TITokenField uses a special string (kTextEmpty) to indicate an empty
text field. This is used to manage token layout, but breaks the default
placeholder text behavior of UITextField, since the text property is
never set to nil.  This change adds a custom placeholderLabel that
appears when the text == kTextEmpty and tokens.count == 0.